Testing

+++ index.js
@@ -5,6 +5,7 @@ require('./mongo')
 const express = require('express')
 const app = express()
 const cors = require('cors')
+const mongoose = require('mongoose')
 const Note = require('./models/Note')
 const notFound = require('./middleware/notFound')
 const handleErrors = require('./middleware/handleErrors')
@@ -17,8 +18,9 @@ app.get('/', (request, response) => {
   response.send('<h1>Hello World!</h1>')
 })
 
-app.get('/api/notes', (request, response) => {
-  Note.find({}).then(notes => response.json(notes))
+app.get('/api/notes', async (request, response) => {
+  const notes = await Note.find({})
+  response.json(notes)
 })
 
 app.get('/api/notes/:id', (request, response, next) => {
@@ -28,7 +30,7 @@ app.get('/api/notes/:id', (request, response, next) => {
     if (note) response.json(note)
     else response.status(404).end()
   })
-  .catch(error => next(error))
+    .catch(error => next(error))
 })
 
 app.put('/api/notes/:id', (request, response, next) => {
@@ -41,19 +43,30 @@ app.put('/api/notes/:id', (request, response, next) => {
   }
 
   Note.findByIdAndUpdate(id, newNoteInfo, {new: true})
-  .then(result => response.json(result))
-  .catch(error => next(error))
-})
-
-app.delete('/api/notes/:id', (request, response) => {
-  const { id } = request.params
-
-  Note.findByIdAndDelete(id)
-    .then(() => response.status(204).end())
+    .then(result => response.json(result))
     .catch(error => next(error))
 })
 
-app.post('/api/notes', (request, response) => {
+app.delete('/api/notes/:id', async (request, response, next) => {
+  const { id } = request.params
+
+  // Verifica si el ID es un ObjectId válido
+  if (!mongoose.Types.ObjectId.isValid(id)) {
+    return response.status(400).json({ error: 'malformatted id' })
+  }
+
+  try {
+    const result = await Note.findByIdAndDelete(id)
+    if (!result) {
+      return response.status(404).json({ error: 'Note not found' })
+    }
+    response.status(204).end()
+  } catch (error) {
+    next(error) // Manejo de otros errores
+  }
+})
+
+app.post('/api/notes', async (request, response, next) => {
   const note = request.body
 
   if (!note || !note.content) {
@@ -68,17 +81,24 @@ app.post('/api/notes', (request, response) => {
     date: new Date(),
   })
 
-  newNote.save().then(savedNote => {
+  /* newNote.save().then(savedNote => {
     response.json(savedNote)
-  })
+  }) */
 
-  response.json(newNote)
+  try {
+    const savedNote = await newNote.save()
+    response.json(savedNote)
+  } catch (error) {
+    next(error)
+  }
 })
 
 app.use(notFound)
 app.use(handleErrors)
 
 const PORT = process.env.PORT
-app.listen(PORT, () => {
+const server = app.listen(PORT, () => {
   console.log(`Server running on port ${PORT}`)
 })
+
+module.exports = { app, server }
